requirements.txt
llama-cpp-python
huggingface_hub
python-dotenv